Struct master_node::Settings

pub struct Settings {
    pub index_dir: PathBuf,
    pub port: usize,
}

Fields
index_dir: PathBuf port: usize
Methods

impl Settings

pub fn from_env() -> Result<Settings, Error>

Load the Settings from the environment.

This will use the SHOUTY_SNAKE_CASE version of each element's name, falling back to the defaults if not present.

Examples

If you wanted to set the port to 1337 and index_dir to /, you would use something like this:

$ PORT=1337 INDEX_DIR=/ cargo run